requirements.txt
pydantic
pymysql
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart